simhospital-python/pyproject.toml
[project]
name = "simhospital"
version = "1.0.0"
description = "Hospital simulation engine producing HL7v2 message streams"
requires-python = ">=3.12"
dependencies = [
    "hl7>=0.4.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simhospital"]

[tool.pytest.ini_options]
testpaths = ["tests"]
